# doc/api.yaml
openapi: 3.0.0
servers:
  - url: "http://localhost:3000" #Api runs locally in port 3000
info:
  title: WASAPhoto
  description: "WASAPhoto API documentation"
  version: 1.0.0
components:
  securitySchemes:  #Security of API, for authentication use Bearer where token represented in JWT Format
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
  /session:    #1)path (URL-address) where we can address to API
    post:      #Since we create new object (new session of user)
      tags:
        - Authentication
      summary: "Logs in the user"
      description: "Authenticates the user and returns an access token. If the user does not exist, it will be created."
      operationId: doLogin
      requestBody:  #description of data what API expect from request
        description: "User credentials object"
        required: true
        content:
          application/json: #data should be in JSON format
            schema:
              type: object  #in body of request we send object
              description: "User credentials for login"
              properties: #list of parameters that we can send
                username:
                  type: string
                  description: "The user's username"
                  example: "Maria"
                  minLength: 3   #minimal length is 3 symbols
                  maxLength: 32  # maximum length is 32 symbols
                  pattern: "^[a-zA-Z0-9_]+$"  #restriction for the format ,means we can use only letter a-z,A-Z,numbers from 0-9 and _
              required:
                - username
      responses:
        '200':
          description: "Successful login"
          content:
            application/json:
              schema:
                type: object
                description: "Response containing the access token and user details"
                properties:
                  token:
                    type: string
                    description: "JWT access token"
                    example: "eyJhbGciOiJIUzI1NiIsInR..."
                    minLength: 1
                    maxLength: 512
                    pattern: "^[A-Za-z0-9\\-_.]+$"
                  id:
                    type: integer
                    format: int64
                    description: "User ID"
                    example: 123
                  username:
                    type: string
                    description: "The user's username"
                    example: "Maria"
                    minLength: 3
                    maxLength: 32
                    pattern: "^[a-zA-Z0-9_]+$"
                required:
                  - token
                  - id
                  - username
        '400':
          description: "Bad request (invalid input)"
          content:
            application/json:
              schema:
                type: object
                description: "Error response"
                properties:
                  error:
                    type: string
                    description: "Error message"
                    example: "Invalid request body"
                    minLength: 1
                    maxLength: 256
                    pattern: ".*"
                required:
                  - error
...
  /group: #7) creating new group
    post:
      tags:
        - Groups
      summary: "Creates a new group"
      description: "Allows a user to create a new group with selected members and an optional profile photo."
      operationId: createGroup
      security:
        - BearerAuth: []
      requestBody:
        description: "Group creation form data"
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              description: "Data for creating a new group"
              properties:
                groupName:
                  type: string
                  description: "Name of the group"
                  example: "Eco Warriors"
                  minLength: 1
                  maxLength: 100
                  pattern: "^[\\w\\s]+$"
                selectedUsers:
                  type: string
                  description: "Comma-separated list of user IDs to be added to the group"
                  example: "123,456,789"
                  minLength: 1
                  maxLength: 256
                  pattern: "^(\\d+(,\\d+)*)?$"
                profilePhoto:
                  type: string
                  format: binary
                  description: "Optional profile picture for the group"
                  minLength: 1
                  maxLength: 256
              required:
                - groupName
                - selectedUsers
      responses:
        '201':
          description: "Group created successfully"
          content:
            application/json:
              schema:
                type: object
                description: "Response indicating successful group creation"
                properties:
                  success:
                    type: boolean
                    description: "Indicates if the group was created successfully"
                    example: true
                required:
                  - success
        '400':
          description: "Bad request (missing required fields or invalid data)"
          content:
            application/json:
              schema:
                type: object
                description: "Error response"
                properties:
                  error:
                    type: string
                    description: "Error message"
                    example: "Group name is required"
                    minLength: 1
                    maxLength: 256
                    pattern: ".*"
                required:
                  - error
        '401':
          description: "Unauthorized (user not authenticated)"
          content:
            application/json:
              schema:
                type: object
                description: "Unauthorized error response"
                properties:
                  error:
                    type: string
                    description: "Error message"
                    example: "Unauthorized access"
                    minLength: 1
                    maxLength: 256
                    pattern: ".*"
                required:
                  - error
  /group/users: #8)Adding Users to a Group
    post:
      tags:
        - Groups
      summary: "Adds users to a group"
      description: "Allows adding multiple users to an existing group."
      operationId: addToGroup
      security:
        - BearerAuth: []
      requestBody:
        description: "Users to be added to a group"
        required: true
        content:
          application/json:
            schema:
              type: object
              description: "Request object for adding users to a group"
              properties:
                groupId:
                  type: integer
                  format: int64
                  description: "ID of the group"
                  example: 101
                userIds:
                  type: array
                  description: "List of user IDs to be added"
                  minItems: 1
                  maxItems: 1000
                  items:
                    type: integer
                    format: int64
              required:
                - groupId
                - userIds
      responses:
        '200':
          description: "Users added to the group successfully"
          content:
            application/json:
              schema:
                type: object
                description: "Response indicating users were added"
                properties:
                  success:
                    type: boolean
                    description: "Indicates if the users were added successfully"
                    example: true
                required:
                  - success
        '400':
          description: "Bad request (invalid request body)"
          content:
            application/json:
              schema:
                type: object
                description: "Error response"
                properties:
                  error:
                    type: string
                    description: "Error message"
                    example: "Invalid JSON format"
                    minLength: 1
                    maxLength: 256
                    pattern: ".*"
                required:
                  - error
        '401':
          description: "Unauthorized (user not authenticated)"
          content:
            application/json:
              schema:
                type: object
                description: "Unauthorized error response"
                properties:
                  error:
                    type: string
                    description: "Error message"
                    example: "Unauthorized access"
                    minLength: 1
                    maxLength: 256
                    pattern: ".*"
                required:
                  - error
...
  /message:
    post:
      tags:
        - Messages
      summary: "Delete a message"
      description: "Allows a user to delete a specific message by its ID."
      operationId: deleteMessage
      security:
        - BearerAuth: []
      requestBody:
        description: "Message delete request"
        required: true
        content:
          application/json:
            schema:
              type: object
              description: "Request object for deleting a message"
              properties:
                messageId:
                  type: integer
                  format: int64
                  description: "ID of the message to be deleted"
                  example: 789
              required:
                - messageId
      responses:
        '200':
          description: "Message deleted successfully"
          content:
            application/json:
              schema:
                type: object
                description: "Response indicating successful deletion"
                properties:
                  success:
                    type: boolean
                    description: "Indicates if the deletion was successful"
                    example: true
                required:
                  - success
        '400':
          description: "Bad request (failed to delete message)"
          content:
            application/json:
              schema:
                type: object
                description: "Error response"
                properties:
                  error:
                    type: string
                    description: "Error message"
                    example: "Failed to delete message"
                    minLength: 1
                    maxLength: 256
                    pattern: ".*"
                required:
                  - error
        '401':
          description: "Unauthorized (user not authenticated)"
          content:
            application/json:
              schema:
                type: object
                description: "Unauthorized error response"
                properties:
                  error:
                    type: string
                    description: "Error message"
                    example: "Unauthorized access"
                    minLength: 1
                    maxLength: 256
                    pattern: ".*"
                required:
                  - error
...
  /message/comment:
    post:
      tags:
        - Messages
      summary: "Add a comment to a message"
      description: "Allows a user to add a comment to an existing message."
      operationId: commentMessage
      security:
        - BearerAuth: []
      requestBody:
        description: "Comment details"
        required: true
        content:
          application/json:
            schema:
              type: object
              description: "Request object for adding a comment to a message"
              properties:
                messageId:
                  type: integer
                  format: int64
                  description: "ID of the message to comment on"
                  example: 12345
                comment:
                  type: string
                  description: "Comment text"
                  example: "This is a great message!"
                  minLength: 1
                  maxLength: 500
                  pattern: "^[\\s\\S]+$"
              required:
                - messageId
                - comment
      responses:
        '200':
          description: "Comment added successfully"
          content:
            application/json:
              schema:
                type: object
                description: "Response containing details of the added comment"
                properties:
                  success:
                    type: boolean
                    description: "Indicates if the comment was added successfully"
                    example: true
                  commentId:
                    type: integer
                    format: int64
                    description: "ID of the newly added comment"
                    example: 67890
                required:
                  - success
                  - commentId
        '400':
          description: "Bad request (invalid input)"
          content:
            application/json:
              schema:
                type: object
                description: "Error response"
                properties:
                  error:
                    type: string
                    description: "Error message"
                    example: "Invalid request body"
                    minLength: 1
                    maxLength: 256
                    pattern: ".*"
                required:
                  - error
        '401':
          description: "Unauthorized (user not authenticated)"
          content:
            application/json:
              schema:
                type: object
                description: "Unauthorized error response"
                properties:
                  error:
                    type: string
                    description: "Error message"
                    example: "Unauthorized access"
                    minLength: 1
                    maxLength: 256
                    pattern: ".*"
                required:
                  - error
...